Refactor notebook tokens
