+++ packages/notebook/src/index.ts
@@ -14,3 +14,4 @@ export * from './widget';
 export * from './widgetfactory';
 export * from './modestatus';
 export * from './truststatus';
+export * from './tokens';
